fixed active item to reflect which page is active
